kexec: bypass check on interp program header for ppc64le zImage
When the elf information is built from the executable, kexec checks
for an interp elf program header and fails if it exists. The following
patch adds a new service machine_check_elf_phdr_interp() which lets
platforms eventually bypass this check.

This is needed for the zImage boot wrapper on PPC 64bit little endian.
The executable is linked as position independent and has a interp elf
program header.

An alternative to this patch would be to remove the whole test on the
interp elf program header as kexec seems to be ignoring the information
after all.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>