ppc64: detect zImage files and load the uncompressed vmlinux
The zImage* executable files on ppc64 use a special section called
".kernel:vmlinux.strip" which contains a compressed vmlinux executable
file.

This patch adds a service to detect and unzip such a section in a
malloc'ed buffer. The buffer is then used in elf_ppc64_load() to
load the new vmlinux.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>